Moved Shib migration to its own folder and broke into two scripts
